// src/store/module/PassengerTrips/types.ts
export enum PassengerTripsTypes {
    GET_PASSENGER_TRIPS_REQUEST = 'passengerTrip/GET_PASSENGER_TRIPS_REQUEST',
    GET_PASSENGER_TRIPS_SUCCESS = 'passengerTrip/GET_PASSENGER_TRIPS_SUCCESS',
    GET_PASSENGER_TRIPS_FINISHED_REQUEST = 'passengerTrip/GET_PASSENGER_TRIPS_FINISHED_REQUEST',
    GET_PASSENGER_TRIPS_FINISHED_SUCCESS = 'passengerTrip/GET_PASSENGER_TRIPS_FINISHED_SUCCESS',
    GET_PASSENGER_TRIPS_IN_PROGRESS_REQUEST = 'passengerTrip/GET_PASSENGER_TRIPS_IN_PROGRESS_REQUEST',
    GET_PASSENGER_TRIPS_IN_PROGRESS_SUCCESS = 'passengerTrip/GET_PASSENGER_TRIPS_IN_PROGRESS_SUCCESS',
    GET_PASSENGER_TRIPS_CANCELED_REQUEST = 'passengerTrip/GET_PASSENGER_TRIPS_CANCELED_REQUEST',
    GET_PASSENGER_TRIPS_CANCELED_SUCCESS = 'passengerTrip/GET_PASSENGER_TRIPS_CANCELED_SUCCESS',
}